Break home and contact page into components
